% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_cosine.R
\name{dist_cosine}
\alias{dist_cosine}
\title{Calculate Cosine Similarity}
\usage{
dist_cosine(x, y = NULL)
}
\arguments{
\item{x}{a vector of embeddings, a list of embeddings, or a matrix of
embeddings (one row per embedding vector)}

\item{y}{optional second element, same structure as `x`}
}
\value{
the cosine similarity value (higher values mean similar vectors).
If `x` or `y` are a matrix or list while the other is a single vector,
the function returns a vector of the same length as the list or the number
of rows.
If `y` is NULL, a cosine similarity matrix is returned of values of `x`.
}
\description{
Note that this is a similarity metric (higher values indicate similar
vectors). The underlying simsimd rust crate calculates 1 - cosine.
}
\examples{
dist_cosine(c(1, 2, 3), c(4, 5, 6))

set.seed(123)
N <- 10 # number of embeddings
emb1 <- rnorm(N)
emb_list_3 <- lapply(seq(3), function(x) rnorm(N))
emb_mat_3 <- do.call(rbind, emb_list_3)

# we get 3 rows, one for each element,
# and 10 columns, one for each embedding dimension
dim(emb_mat_3)

# single element - list => vector of similarities
dist_cosine(emb1, emb_list_3)
# arguments can be swapped
dist_cosine(emb_list_3, emb1)

# single element - matrix => vector of similarities
dist_cosine(emb1, emb_mat_3)
# arguments can be swapped
dist_cosine(emb_mat_3, emb1)

# list only results in NxN matrix
dist_cosine(emb_list_3)
# matrix only results in NxN matrix
dist_cosine(emb_mat_3)

# two lists or matrices
emb_mat_5 <- matrix(rnorm(5 * N), ncol = N)

# 3x5 elements results in a 3x5 matrix
dist_cosine(emb_mat_3, emb_mat_5)
}
