% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_capabilities.R
\name{get_capabilities}
\alias{get_capabilities}
\title{Report Current Hardware Capabilities}
\usage{
get_capabilities()
}
\value{
a named vector of capabilities
}
\description{
See also <https://docs.rs/simsimd/latest/simsimd/capabilities/index.html>
}
\examples{
get_capabilities()
}
